updated wallet money to be ulong type

+++ Common/Scripts/Wallet.cs
@@ -1,8 +1,7 @@
 using Godot;
-using System;
 
 [GlobalClass]
 public partial class Wallet : Resource
 {
-    [Export] public int Money { get; set; }
+    [Export] public ulong Money { get; set; }
 }
